build-sys: support newer gee version
"Libgee version numbers are a little confusing. In libgee 0.5 and 0.6,
the package name was gee-1.0, identified by a file gee-1.0.pc. In
libgee 0.7, the package name changed to gee-0.8 (since the library
changed incompatibly)."

Since systemadm seems to work fine with newer libgee (found in Fedora 17),
let's check for both versions, starting with the newer one.

Between 1.0 and 0.8 there seems to have been 0.7, but since I don't
have one at hand, I'm not adding a check for it. If necessary,
configure.ac can be extended to check for that version too.